Move cells to Main class
